cmd added in setup

+++ setup.py
@@ -3,7 +3,10 @@ import DbManager
 
 if __name__ == '__main__':
     if(not os.path.isdir('./static/image')): os.mkdir('./static/image')
-    if(os.path.isfile('./Data.db')): os.remove('./Data.db')
-    DbMan = DbManager.DbManager()
-    DbMan.makeDb()
-    DbMan.closeDb()
+    cmd = int(input("Del Db? yes->1 no->0"))
+    if cmd == 1:
+        if(os.path.isfile('./Data.db')): os.remove('./Data.db')
+        DbMan = DbManager.DbManager()
+        DbMan.makeDb()
+        DbMan.closeDb()
+
